+++ src/main/java/com/team01/deokhugam/book/BookMapper.java
@@ -1,6 +1,7 @@
 package com.team01.deokhugam.book;
 
 import com.team01.deokhugam.book.dto.BookDto;
+import com.team01.deokhugam.book.entity.Book;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
