+++ src/main/java/com/team01/deokhugam/book/repository/BookRepository.java
@@ -1,10 +1,8 @@
 package com.team01.deokhugam.book.repository;
 
-import com.team01.deokhugam.book.Book;
+import com.team01.deokhugam.book.entity.Book;
 import java.util.Optional;
 import java.util.UUID;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository<Book, UUID>, BookRepositoryQueryDsl {
