+++ src/main/java/com/team01/deokhugam/review/service/ReviewServiceImpl.java
@@ -1,6 +1,6 @@
 package com.team01.deokhugam.review.service;
 
-import com.team01.deokhugam.book.Book;
+import com.team01.deokhugam.book.entity.Book;
 import com.team01.deokhugam.book.repository.BookRepository;
 import com.team01.deokhugam.global.exception.book.BookNotFoundException;
 import com.team01.deokhugam.review.dto.ReviewCreateRequest;
