+++ src/test/java/com/team01/deokhugam/book/BookServiceTest.java
@@ -13,7 +13,9 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import com.team01.deokhugam.book.dto.BookCreateRequest;
 import com.team01.deokhugam.book.dto.BookDto;
+import com.team01.deokhugam.book.entity.Book;
 import com.team01.deokhugam.book.repository.BookRepository;
+import com.team01.deokhugam.book.service.BookService;
 import com.team01.deokhugam.global.exception.book.BookNotFoundException;
 import com.team01.deokhugam.book.dto.BookUpdateRequest;
 import com.team01.deokhugam.global.exception.book.DuplicatedIsbnException;
